move min-cost-flow rebalancing to a separate package
